Suppression Warning Windows

+++ src/AMD/amd_aat.cpp
@@ -179,7 +179,7 @@ GLOBAL size_t AMD_aat	/* returns nz in A+A' */
 	Info [AMD_NZ] = nz ;
 	Info [AMD_SYMMETRY] = sym ;	    /* symmetry of pattern of A */
 	Info [AMD_NZDIAG] = nzdiag ;	    /* nonzeros on diagonal of A */
-	Info [AMD_NZ_A_PLUS_AT] = nzaat ;   /* nonzeros in A+A' */
+	Info [AMD_NZ_A_PLUS_AT] = (double)nzaat ;   /* nonzeros in A+A' */
     }
 
     return (nzaat) ;
